More configs
Wanted to change where and how the configs worked.
Specifically, the .zshrc file in the home directory sources the .zsh/.zsh file.
The .zshrc file is only used as a configuration file for the main system.

Of course, if one needs to update the config file, they will need to do it manually.